test: notification repository unit test 파일 통합
